refactor: ajusta nome do oauth no swagger docs
