+++ src/docs/swagger/auth/oauth.ts
